Kernel/PCI: Attach SysFS PCI directories references to DeviceIdentifiers
This will be useful when we want to attach PCI graphics adapters with
the "host_device" node in each SysFS graphics adapter directory to a PCI
SysFS device directory with a symlink.
