+++ Kernel/FileSystem/SysFS/Subsystems/Bus/PCI/BusDirectory.h
@@ -6,7 +6,6 @@
 
 #pragma once
 
-#include <Kernel/Bus/PCI/Definitions.h>
 #include <Kernel/FileSystem/SysFS/Component.h>
 
 namespace Kernel {
