+++ Kernel/FileSystem/SysFS/Subsystems/Bus/PCI/DeviceDirectory.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <Kernel/Bus/PCI/Definitions.h>
+#include <Kernel/Bus/PCI/Address.h>
 #include <Kernel/FileSystem/SysFS/Component.h>
 #include <Kernel/KString.h>
